Feat: Make more test friendly with non global http client
In order to enable testing with mock http clients, this commit adds two
new initializers to the AsanaClient struct. They allow passing in an
httpclient::Client instance which means we can pass in mock/stub ones
during testing and prevent real api calls. This changes how the struct
holds onto the default client as must now own the client instead of
holding a static lifetime reference. To avoid breaking the default
behavior, we need to clone the static singular instance so we can have
an owned copy. This means all of the behavior from the static instance
is shared but it is a new copy every time we create AsanaClient.